refactor(lib): split detect-musl-cc and build-outputs out of resolve-setup
resolve-setup.ts was up to 1072 lines after the recent feature work
(enable, musl cc-rs, cache-shutdown-on-idle). Pull two self-contained
sections into their own modules to keep the orchestrator focused on
input parsing / cache-key derivation:

* `src/lib/detect-musl-cc.ts` — `detectMuslCcEnv`,
  `tripleToCcRsSuffix`, the two private PATH helpers, and the
  `MuslCcResolution` / `DetectMuslCcDeps` types (~130 LOC).
* `src/lib/build-outputs.ts` — `buildOutputs`, the pure
  ResolveResult → `$GITHUB_OUTPUT` mapping (~52 LOC).

`resolve-setup.ts` re-exports `detectMuslCcEnv`, the related types,
and `buildOutputs` so existing test imports keep working without any
test edits. No behavior change; full suite stays at 201 pass / 4
skipped.

resolve-setup.ts: 1072 → 898 lines.
